Rework import lookups
Make LookupScope responsible for incorporating imports that are visible
at the point to which given LookupScope instance corresponds to.

A single LookupScope (e.g. corresponding to a class) might need to
incorporate different sets of imports (if imports are inside class
definition and interleaved with other definitions). For that reason,
LookScope has now a method that clones an existing lookup scope but
replaces ImportsLookupScope.

This simplifies import handling that previously was really hard to follow
in its implementation.