Implement completing member list for classes
Implement a very basic template completer that calculates the list of all
members of classes. This completer doesn't work with objects yet.

The test demonstrates resolution of members works across inheritance. It
also demonstrates that dependencies between completers are handled through
an ordered job queue. If a completer is missing a dependency, it explicitly
yields back to job queue with information about the missing dependency.