Implement DefDefCompleter
Implement a basic functionality of a DefDefCompleter that computes a type
of a method. The completer supports just one parameter list and a very
basic form of parameter types:

   def foo(a: ident.select1.select2, ...)

For example, `scala.Int` is supported but `Map[Int, Int]` is not because
we don't support resolving application of polymorphic types yet.

This commit adds scheduling and executing of DefDefCompleters. The current,
implementation schedules all declared members of a class to be
completed right after all the list of members of the class is completed.
For experimentation, the current scheduling of completers is fine but not
typechecking of all Scala code be scheduled in that way. When nested
classes are involved, one can create a cycle when in order to compute
member list one needs to understand some of member signatures.

Lastly, benchmarks has been added for method completion. Interestingly,
we can still achieve 1700 ops/sec of completing all signatures in
10k.scala. This translates to 17 million lines of code processed per
second for completing signatures.