Add support for completing modules (objects)
Implement support for completing member list of objects by using
TemplateMemberListcompleter. This completer works with classes so each
module symbol gets a corresponding class symbol that actually holds info
about members. This design follows the implementation scheme of both
scalac and dotty. We might revisit the decision of creating a synthetic
class symbol for modules in the future if better design idea emerges.