Completing vals and polymorphic class parents
Add support for completing signature (a result type) of a val. The
completer is a simplified version of a def completer.

Add support for polymorphic class parent (classes that have type
parameters). This required introduction of processing type trees. Only
support for AppliedTypeTree has been introduced so C[Int, String] is
supported as a parent.

There's a test that shows polymorphic classes are handled now but inherited
members do not get type parameters substituted with available type
arguments.