Implement substitution of type parameters
Implement substitution of type parameters with type arguments for classes.
Let's consider the following scenario

class B extends A[C]
class A[T] { val a: T }
class C

When we look at the val `a` from `B` we should see val's result type to
be `C`. For that to be the case, we need to substitute the type parameter
`T` with type argument `C`.

The substitution logic is implemented by `substituteTypeArgs` methods.
Inherited members with substituted type arguments are entered in
TemplateMemberListCompleter. For this to work, we need parents to have
members with completed signatures. This breaks previous separation of
completing templates and members in two distinct steps. For that reason,
this commit unifies job queues into one completer job queue.

The performance is not affected in any noticeable way by this commit
we still maintain the following performance for completing-related
benchmark

[info] Benchmark                                             (filePath)   Mode  Cnt     Score    Error  Units
[info] BenchmarkEnter.completeMemberSigs  sample-files/10k.scala.ignore  thrpt   60  1706.795 ± 16.990  ops/s