Overhaul module completion and lookups
Perform lookups using symbol's info instead of symbol's declarations so
we see all members (including inherited ones) of a symbol.

The changes were rather significant:

  * Type has a `lookup` method now
  * all places go through info for lookup calls

Apart from that, module completion was refactored. We have a completer for
module class symbol that carries the actual list of declarations and
members. Then, we have a completer for module symbol that just assigns
a type that points to module class info.