Add BenchmarkScalap
Add jmh benchmarking code that measures performance of:

  * entering symbols into symbol table for scalap sources
  * completing member signatures of all members for scalap sources

The current performance numbers are:

[info] Benchmark                                    Mode  Cnt      Score     Error  Units
[info] BenchmarkScalap.completeMemberSigs          thrpt   60   2058.282 ±  15.474  ops/s
[info] BenchmarkScalap.enter                       thrpt   60  12369.898 ± 285.836  ops/s
[info] BenchmarkScalap.parse                       thrpt   60     88.459 ±   0.683  ops/s
[info] BenchmarkScalap.parseAndCompleteMemberSigs  thrpt   60     83.209 ±   0.798  ops/s

The scalap project has 2171 lines which yields 191k lines of code parsed
per second. The completeMemberSigs benchmark focused on measuring the
performance of building the symbol table and completing member signatures
of all entered symbols. For this benchmark we get the performance of
4.4 million lines of code processed per second.

It is a bit surprising that building a symbol table and completing symbol
types is over 23 times faster than parsing Scala files.