Construct the DAG in Tarjan's algorithm
Add implementation of constructing the edges of the DAG formed by Strongly
Connected Components calculated by the Tarjan's algorithm.

The cool thing about the implementation that it's a modest modification
of the algorithm. I introduced just one additional stack (for components)
and a few hashmaps for mappings. The algorithm is still a single pass and
is probably the most efficient way of calculating the DAG formed by SCCs.