Add TarjanSCC.longestPath implementation
Add implementation of the algorithm for finding a longest path in a DAG
returned by Tarjan's algorithm. There can be one than more longest path
and arbitrary one is returned by the algorithm.

The code is written with high performance as its main goal so it's rather
low-level and ugly. I don't have a good intuition which high-order methods
on Scala collections are cheap to use so I'm not using any of them.