Implement dependency extraction between Symbols
Add an implementation of `DependenciesExtraction` that walks symbol table
and extracts all dependencies between symbols by inspecting outline types.

Dependencies are tracked at class level. I don't think a more fine
granurality is that useful but the code would be easy to modify if a
more fine or more coarse granurality is needed. As usual, the code has
been written with paranoid approach to achieving high performance:
everything that is not proved to be cheap is assumed to be expensive.
For most part, that means I drop to "Java in Scala" style of coding.

Main calls dependency extraction, then collapses the graph using
Tarjan's algorithm for computing Strongly Connected Components. Lastly,
Main finds the longest path in dependency graph between components and
prints it.

It turns out that for scalap, the longest dependency path is 10 elements
long. Full analysis of scalap, including dependency analysis finishes
in 453 miliseconds.