Refactor nested while loops in DefDefCompleter
Refactor two loops into nested `map` calls.

Benchmarks shows that this is actually faster than a while
loop accumulating into Java list and then a copy to a Scala
List:

perf before this refactor:

[info] Benchmark                            Mode  Cnt      Score    Error  Units
[info] BenchmarkScalap.completeMemberSigs  thrpt   60   2432.630 ± 27.027  ops/s
[info] BenchmarkScalap.enter               thrpt   60  16302.149 ± 85.146  ops/s

perf with this refactor:

[info] Benchmark                            Mode  Cnt      Score     Error  Units
[info] BenchmarkScalap.completeMemberSigs  thrpt   60   2693.014 ±  27.008  ops/s
[info] BenchmarkScalap.enter               thrpt   60  16434.142 ± 315.295  ops/s

This is rather significant jump in performance. It makes sense, though.

The `List.map` operation is implemented efficiently in Scala because it can build
the single linked list in a single pass thanks to mutable references that are
accessible just to the collection operations. In Scala 2.12 `map` is also inlined
as confirmed by the inliner log:

Inline into kentuckymule/core/Enter$DefDefCompleter.complete: inlined scala/collection/immutable/List.map. Before: 203 ins, inlined: 72 ins.

The conclusion is: if we want to keep using Scala list, then using its native
operations is the best choice for performance.