Add lookup in the scala package.
Lookup scala package right before deferring to the _root_ package.

Add support for lookups in Predef

Add a bit hacky support for looking up in Predef object. The spec says
that it's implicitly imported but we're making PRedef simply a part
of lookup scope chain which is slightly off. I hope to fix that later
when I identify this to be a problem.

Benchmarks suggest a negligible negative performance cost: around 2%
slowdown in the BenchmarkScalap.completeMemberSigs.

The latest results (on my old laptop) are:

[info] # Run complete. Total time: 00:04:38
[info] Benchmark                            Mode  Cnt      Score    Error  Units
[info] BenchmarkScalap.completeMemberSigs  thrpt   60   2124.040 ± 17.542  ops/s
[info] BenchmarkScalap.enter               thrpt   60  12078.164 ± 68.100  ops/s