The `processJobQueue` returns `CompleterStats`
The CompleterStats stored both number of processed jobs and dependency
misses. It will get expanded in the future with additional high level
stats on completers' performance.

The immediate use case for these stats is to get some sense how often
the completers get retried due to a dependency miss. An idea for
the future: count how many completers get retried due to the same
missing dependency. That might suggest a presorting of the queue
might be a good optimization strategy. For example, one can imagine
that most of other completers should be ran _after_ all/most packages
are already completed because everything else depends on packages.