Fix scoping for the empty package
Scala's parser has a weird corner case when both an empty package and
regular one are involved in the same compilation unit:

// A.scala
class A
package foo {
  class B
}

is expanded during parsing into:

package <empty> {
  class A
  package foo {
    class B
  }
}

However, one would expect the actual ast representation to be:

package <empty> {
  class A
}
package foo {
  class B
}

I believe `foo` is put into the empty package to preserve the property
that a compilation unit has just one root ast node. Surprisingly, both
the scalac and dottyc handle the scoping properly when the asts are
nested in this weird fashion. For example, in scalac `B` won't see the
`A` class even if it seems like it should when one looks at the nesting
structure. I couldn't track down where the special logic that is
responsible for special casing the empty package and ignoring the nesting
structure.

Kentucky Mule didn't have any special rule until now for handling empty
packages so it would get the nesting wrong. Now, we normalize the trees
so empty package contains just class and object declarations, whereas
packages are pulled out of the empty package.

Added a test that covers that logic.