Create and enter the `emptyPackage` symbol in Definitions
Always create and enter the `emptyPackage` symbol in Definitions. The
motivation is that we can refer to this symbol from the code in Enter so
we want to have a canonical reference to it.

The unconditional creation of this symbol affects the EnterTest because of
an additional descendant path introduced. I refactored the
`descendantPaths` to filter out `rootPackage :: emptyPackage :: Nil` path.
This required introduction of an implicit `context` parameter.

As a result of passing context, I could introduce a helper
`descendantPathsFromRoot` method and switch all tests to using it. I also
removed repeated `implicit val context = ctx` declarations by making
the outer val implicit.