Sprinkle a few missing cases: Apply, Super, This and SingletonTypeTree
Strip away `Apply` in a context where an applied type is expected:
the `Apply` is a constructor call as in

  class Foo extends Bar[T]()(t)

and we're interested just in the `Bar[T]` part.

Singleton type `p.x.type` is where `p` is a path and `x` is a member,
is resolved to result type of `x` (e.g. a type of a val)

Super and This are simply stripped down and so either `C.this` or `C.super`
are resolved to `C`.

See the corresponding tests for details.