Add support for type aliases
Add support for both non-parametric and parametric type aliases.
This is a large commit so it comes with a larger explanation.

Let's start with an example:

abstract class Base[A, B] {
  val b: B
}
class Foo
class Bar
class A {
  type Flip[X, Y] = Base[Y, X]
  type Result = Flip[Foo, Bar]
}

If we're looking at `Result.b` which is `Base.b` once dealiased, the result
type of `b` should be `Foo` because type aliases go through `Flip`.

This commit is about making that happen.

Lets me highlight a few changes:

  * LookupTypeDefScope has been introduced so the right hand side of type
    alias can lookup type parameters of a type (e.g. Flip's rhs)
  * TypeAliasCompleter has been added that performs the actual completion
    of a type; it's worth noting that the completer does not perform any
    dealiasing during the completion so if rhs points at an alias type,
    that information will be preserved
  * TemplateMemberListCompleter now supports completing classes that refer
    to type aliases in their parents, this is implemented using
    AppliedTypeMemberDerivation

The AppliedTypeMemberDerivation is subtle and deserves its own paragraph
to explain. Let's extend our example to be:

class A {
  type Flip[X, Y] = Base[Y, X]
  type Result = Flip[Foo, Bar]
  class B extends Flip[Foo, Bar]
}

we want to now the type of `B.b`. Kentucky Mule copies members from its
parents and performs the right type substitutions which is a simplified
version of the As Seen From algorithm.

The AppliedTypeMemberDerivation class precomputes all information required
to go from an applied type like `Flip[Foo, Bar]` to a member of a class
`Flip` dealiases to with all types substituted according to the arguments
threaded through type applications.

The crucial property of the AppliedTypeMemberDerivation is that its
computed once for a given parent and applied repeatedly for all inherited
members. The cost of applying AppliedTypeMemberDerivation is linear with
respect to the number of type parameters of the underlying class where a
member is defined.

So if `Base` has `n` members and `m` type parameters, the cost of computing
all inherited members from `Base` is `n*m` + cost of preparing the
AppliedTypeMemberDerivation. Check the comments in TypeOps for more
in depth analysis.

The added tests check for both non-parametric and parametric type aliases
that appear as parents of a class. Also, I added a test that is more
narrowly focused on the AppliedTypeMemberDerivation functionality.

Finally, this fairly complex functionality have negligible performance cost:
around 1.5%.

```
sbt:root> ;clean; compile; bench/jmh:run kentuckymule.bench.BenchmarkScalap.completeMemberSigs -f 6
```

Result before this change:

```
[info] Result "kentuckymule.bench.BenchmarkScalap.completeMemberSigs":
[info]   1960.697 ±(99.9%) 13.943 ops/s [Average]
[info]   (min, avg, max) = (1831.083, 1960.697, 2046.041), stdev = 45.266
[info]   CI (99.9%): [1946.754, 1974.639] (assumes normal distribution)
[info] # Run complete. Total time: 00:04:08
[info] Benchmark                            Mode  Cnt     Score    Error  Units
[info] BenchmarkScalap.completeMemberSigs  thrpt  120  1960.697 ± 13.943  ops/s
```

With this change:

```
[info] Result "kentuckymule.bench.BenchmarkScalap.completeMemberSigs":
[info]   1923.250 ±(99.9%) 10.106 ops/s [Average]
[info]   (min, avg, max) = (1824.787, 1923.250, 1988.683), stdev = 32.810
[info]   CI (99.9%): [1913.144, 1933.356] (assumes normal distribution)
[info] # Run complete. Total time: 00:04:08
[info] Benchmark                            Mode  Cnt     Score    Error  Units
[info] BenchmarkScalap.completeMemberSigs  thrpt  120  1923.250 ± 10.106  ops/s
```