Fix for forgetting imports across scopes
Fix a bug that would cause some imports to be forgotten.
Let's consider this example:

  package foo
  import A.B
  object A {
    object B {
      class C
    }
    class B
  }
  class X {
     import B.C
     def c: C
     def b: B
  }

In the lookup scope of `X`, the `import A.B` from the parent scope
was dropped due to a bug. The bug was coming from my confusion about
how to implement lookup in imports collected within a scope,
e.g. a class.

The fix is actually pretty straightforward: remove `replaceImports`
method from LookupScope that always looked like a code smell.
Ironically, `replaceImports` did what it said: replaced imports so
they would get dropped.

Lookup that takes into account imports is now implemented via
`ImportsLookupScope` that combines lookup from `ImportsLookup`
and lookup in a parent scope.

I considered turning `ImportsLookup` into a `LookupScop` that would
be simply: lookup in imports and forward to parent. However, this
is not possible because imports have lower priority than members declared
in their enclosing scope. So the correct order of lookups is:

1. check members
2. check imports
3. delegate to parent scope

The changes from this patch have no negative effect on the performance:

[info] Result "kentuckymule.bench.BenchmarkScalap.completeMemberSigs":
[info]   1926.658 ±(99.9%) 11.912 ops/s [Average]
[info]   (min, avg, max) = (1846.320, 1926.658, 2025.072), stdev = 38.672
[info]   CI (99.9%): [1914.746, 1938.570] (assumes normal distribution)
[info] # Run complete. Total time: 00:04:08
[info] Benchmark                            Mode  Cnt     Score    Error  Units
[info] BenchmarkScalap.completeMemberSigs  thrpt  120  1926.658 ± 11.912  ops/s