Add many more java std lib stub classes
This patch adds many more stub classes for Java std library.
It's surprising how many dependencies Scala's std library has on
Java std library. A big offender is definitely the
`scala.concurrent` package that has a deep integration with JDK's
concurrency primitives. This is actually not that surprising.

In addition to a large number of JDK's classes, there are a few
Java classes that are defined in scala's std library. Kentucky
Mule doesn't parse Java sources so these classes have to be stubbed
as well. They live in the `scala`, `scala.math` packages.