Comment out some non-essential java imports
For example, imports that pull in members of Integer. They're boring cases
that we gain nothing from implementing stubs for.

Another example is java.security.AccessControlException that is referred
to just from the method body so commenting out the import doesn't change
anything about resolving types in signatures but saves me a bounch of
work to stub it.