Queue refactor: introduce CompletionJob and TypeAssigner
Refactor `CompletersQueue.processJobQueue`:

  1. Move type assignments logic to TypeAssigner
  2. Introduce CompletionJob that encapsulates calling a completer

The TypeAssigner encapsulates the actual mutation of the symbol table
where we set symbol's info after its completer finishes.

The introduction of the CompletionJob is the first step towards
abstracting over the implementation details of jobs the queue processes.
At the end of that process, the queue won't have any code specific to
symbol table and will be focused only on scheduling jobs and handling
their dependencies.