Do not mutate queue in scheduleMembersCompletion
Instead of mutating the queue directly in the scheduleMembersCompletion
method, return an array list of spawned jobs.

Surprisingly, this refator has a large performance impact. Before, the
performance was 1930ops/s and now:

[info] Result "kentuckymule.bench.BenchmarkScalap.completeMemberSigs":
[info]   1851.952 ±(99.9%) 7.788 ops/s [Average]
[info]   (min, avg, max) = (1794.082, 1851.952, 1914.277), stdev = 25.286
[info]   CI (99.9%): [1844.163, 1859.740] (assumes normal distribution)
[info] # Run complete. Total time: 00:04:08
[info] Benchmark                            Mode  Cnt     Score   Error  Units
[info] BenchmarkScalap.completeMemberSigs  thrpt  120  1851.952 ± 7.788  ops/s

That's a 4% performance drop! I don't understand why this refactor would
have such a large performance penalty.

We're doing this refactor to make it possible for CompletionJob to
encapsulate all completer-specific logic so the queue can be generalized
to arbitrary tasks. This will enable easier testing of queue's properties
once more advanced features like cycle detection are added.