Implement CollectingPendingJobsQueueStrategy
Add an implemention supporting CollectingPendingJobsQueueStrategy. This
strategy collects all pending jobs in a side queue and releases all
of them at once when the blocking job is complete. The side queue is
maintained in the queue store associated with the blocking job.

Adding this implementation doesn't have any negative effect on performance:

[info] # Run complete. Total time: 00:04:08
[info] Benchmark                            Mode  Cnt     Score   Error  Units
[info] BenchmarkScalap.completeMemberSigs  thrpt  120  1860.166 ± 9.453  ops/s