Fixes to CollectingPendingJobsQueueStrategy implementation
A number of fixes of the implementation of collecting pending
jobs queue strategy. They're all centered around one idea:

  *** there should be one instance of a completion job for a given
  *** completer

This is important property because we track blocking on other jobs and the
definition whether job is blocking depends on completer's logic. We want
1-1 correspondance between completion jobs and completers. This property
will be crucial for detection of cycles.

This is achieved by the introduction of the mutable field `completionJob`
in the completer.

All other changes in this commit are natural consequence of implementing
the correspondance between completers and completion jobs.

The CollectingPendingJobsQueueStrategy
is enabled by default now. The performance penalty is around 40ops/s:

[info] # Run complete. Total time: 00:04:08
[info] Benchmark                            Mode  Cnt     Score    Error  Units
[info] BenchmarkScalap.completeMemberSigs  thrpt  120  1810.713 ± 10.303  ops/s