Implement job dependency cycle detection in JobQueue
This commit adds cycle detection between jobs executed in a JobQueue. The
implementation is only done for CollectingPendingJobsQueueStrategy.

The cycle detection relies on the property of
CollectingPendingJobsQueueStrategy for pending jobs. Cycle happens when
all jobs has been tried (the queue is empty) and some jobs are left in
the pending state. Once queue is empty, the jobs in pending state are all
part of a cycle, not necessarily the same one, though.

To detect a cycle, we need to check for two conditions:

  * an empty job queue
  * a count of pending jobs

The condition of an empty job queue is pretty subtle to achieve in presence
of cycles. Check comments in `queueIncompleteDependencyJobs` how this is
achieved. The key is to keep track of job's state which is done via
`QueueStore.queued` that this commit adds.

This commits also adds a test for proper cycle detection. The semantics of
the queue are that the queue will execute _all_ jobs that are not part of
any cycle and keep the remaining jobs uncompleted. The test checks for both
of these conditions.

There's no performance penalty, the performance was 1810 ops/s before
this refactor and now is:

[info] # Run complete. Total time: 00:04:08
[info] Benchmark                            Mode  Cnt     Score   Error  Units
[info] BenchmarkScalap.completeMemberSigs  thrpt  120  1811.687 ± 9.295  ops/s