Implement finding a QueueJob dependency cycle
Implement finding a dependency cycle amongst pending jobs for better error
reporting.

The implementation was rather straightforward with all earlier
refactorings and bookkeeping implemented. The main idea is to rerun all
pending jobs to collect their dependencies and then use DFS algorithm to
find a cycle.

Check out the comments on `jobsFindCycle` for details.

Unfortunately, the additional bookkeeping of all pending jobs introduces
a performance hit of 65ops/s. Before this change we had

[info] # Run complete. Total time: 00:04:08
[info] Benchmark                            Mode  Cnt     Score   Error  Units
[info] BenchmarkScalap.completeMemberSigs  thrpt  120  1811.687 ± 9.295  ops/s

and with this change we have:

[info] # Run complete. Total time: 00:04:08
[info] Benchmark                            Mode  Cnt     Score   Error  Units
[info] BenchmarkScalap.completeMemberSigs  thrpt  120  1745.809 ± 9.757  ops/s

It's not clear to me why an additional array list would be so expensive.