JobQueue returns JobQueueResult instead of CompleterStats
Refactor JobQueue.processJobQueue to returns JobQueueResult which is an
ADT that has CompleterStats and JobDependencyCycle as cases. This will
make it possible to return information about detected cycle more cleanly.
Previously, cycles were handled with an exception but cycles are as valid
input as non-cyclic jobs. Cycles come up often in user programs and should
receive a direct, first-class treatment.

Refactored all client code to work with this change.