Add `newMemberLookupScope` to LookupScopeContext
Add `newMemberLookupScope` and make `parentScope` private in
the LookupScopeContext. I opened access to `parentScope` as a quick hack
before to be able to access the enclosing lookup scope without pushing
any new one. Ultimately, this was a confusing move. The purpose of
LookupScopeContext is encapsulation of parent scopes that are chained
together. Breaking that encapsulation led to confusing bugs where I would
grab the wrong LookupScope when I got confused by parent->child
relationship.

Now, we have simply `newMemberLookupScope` which returns a lookup scope
that has allows lookup of members in the whole chain of lookup scopes the
LookupScopeContext instance is aware of.