Add a workaround for cyclic dependencies in package objects
The workaround special cases the lookup of parents of a package object to
perform the lookup only in the package declarations and not in package members.
This breaks the cyclic tie.

The whole situation is an unfortunate consequence of package objects being
ill-specified.

Add a test case that covers this special case and checks its effectivness at
unlocking processing a common pattern of a package object inheritng from a
class in the same package.