Return all cycles from the JobQueue.processJobQueue
Compute all cycles detected in the JobQueue instead of returning just one
arbitrarily (the first one to find). This enables more comprehensive
error reports as demonstrated in Main.

The code pays special attention to this scenario:

If we have a loop like: A <-> B and we have a job C -> A then we expect
that the loop detection algorithm will return only one cycle: A <-> B.

Added a test covering this case.