Add support for self types
Add support for self types that contribute to name lookup within the
class. E.g.

trait Foo[T] {
  class Inner
}

class Bar[T] { self: Foo[T] =>
  def a: Inner
}

In the class `B`, `Inner` resolves to `Foo#Inner` due to self-type refining
the type of `this`. In the implementation we don't route searches through
"this" but bake in the logic to check for a possible self-type into the
`LookupClassTemplateScope`.