Add support for resolving `super` in paths
Add a fairly comprehensive support for resolving the `super` in paths.
It's not a complete support because that requires the full As Seen From
implementation due to super's reliance on the definition of super types.

The approximation is to ignore type applications. The trait qualifier as
in C.super[T] is respected, though. When there's no qualifier specified
all parents are searched in the reverse order. This approximates the
"last declaration wins" rule of mix-in composition.

Add a test case that covers all these scenarios.