Remove Inherited{Def,Val}DefSymbols
Remove the inherited versions of def and val symbols. They were originally
added to support a hacky subset of As Seen From. However, over time they
proved to be too expensive to support: processing scala std library creates
about 3.5 million of these symbols.

In order to not let the ASF functionality bitrot, I introduced the
`asSeenFromThis_slow` method on ClassInfoType. The method is written in a
style that guarantees slow performance. I included `_slow` suffix in its
name to signal that fact. Tests dependent on ASF functionality has been
rewritten to use that method.

Lastly, notes.md include an in-depth discussion of the reasons why these
symbols needed to go.