Do not force import when not necessary
Be more careful about forcing import in order to avoid cycles
in a scenario permitted by Scala. Let's consider this example:

package a {
  package b {
    // this wildcard import is ok: it forces completion of `C` that is not dependent on `A` below
    import C._
    // import A._ causes a cyclic error also in scalac; only importing a specific name
    // avoids the cycle
    import A.X
    object A extends B {
      class X
    }
    object C
  }
  class B
}

Before this change, the `import A.X` would fail due to following
cycle:

  1. `import A.X` needs to query members of the object `A` so
      that object has to be completed
  2. During the completion of the object `A`, we need to resolve
     the type `B` and for that we need to resolve all imports
     in scope

The new logic won't consider all imports but only imports possibly
matching. The possibly matching imports are the ones with a selector
matching the looked up identifier's name.

Wit this commit, the lookup for the type `B` doesn't trigger
the completion of the import `A.X` because the name `B` is not matched
by the selector `X`.

To achieve this effect, this patch introduces:

  * the `possiblyMatches` method that checks whether a completer should
    be forced for a given lookup
  * queries only completed imports symbols

Surpisingly, the additional laziness improves the performance.

Before this patch:

```
[info] # Run complete. Total time: 00:04:08
[info] Benchmark                            Mode  Cnt     Score   Error  Units
[info] BenchmarkScalap.completeMemberSigs  thrpt  120  1736.439 ± 8.539  ops/s
```

and after:

```
[info] # Run complete. Total time: 00:04:08
[info] Benchmark                            Mode  Cnt     Score   Error  Units
[info] BenchmarkScalap.completeMemberSigs  thrpt  120  1864.463 ± 8.079  ops/s
```

The performance of `BenchmarkScalalib.completeMemberSigs` remains unchanged.