Be less strict about self-types
Before this patch, we would compute self-type's signature at the moment
of completing the signature of a template (class/trait/object). This
was too eager, though. Let's consider this example:

```
class A { self: B =>
}
class B extends A
```

While completing the example above, we would run into a cycle:

1. While completing the class `A` we would attempt to also complete
   the variable `self` which would trigger completion of `B`. The
   completion of `B` was forced because all components of ClassInfoType
   were supposed to be completed at the time of creating that type.
2. The completer for the class `B` depends on completion of the class
    `A`

Going through SymRef indirection that allows for returning an Incomplete
Dependency at the _lookup_ time of a self-type and breaks the cycle.