Add a pending `asSeenFromBasic` test
Add a test that shows that the type member refinement is currently not
taken into account when looking up members. Check

https://gist.github.com/gkossakowski/2f6fe6b51d1dd640f462a20e5560bcf2#type-members

for the relationship between ASF and type members.

This test shows that a fairly basic case of as seen from is not properly
supported by Kentucky Mule. Is it a bad news?
It depends. Kentucky Mule is able to process the entire scala library
sources without running into this limitation. How come? This problem pops
up only when path dependent types are used together with type members that
are refined through inheritance. Scala library has some path-dependent
types but they are rare. A proper support might be required to be added but
given how rarely it's required, we can afford for its implementation to
be expensive to run. This is good news.